Feature/use templates menu (#12)
* add header, footer and leftnav

* make page active with highlighting on the link

* add makeAdmin in settings account page and update roleWebClient for api calls

* update settingsController with delete user from superadmin role, update profile search with a query param to return result ignoring searchable in user property

* reformat buttons for settings

* refactored client create code, removed user signup and put in authorization app

* app/src/main/java/me/sonam/authzmanager/controller/admin/clients/settings/SettingsController.java

* add update error message setting controller

* working

* add defaultorganizationid on admin user signup

---------

+++ app/src/main/java/me/sonam/authzmanager/controller/admin/user/DeleteMyDataController.java
@@ -1,9 +1,11 @@
 package me.sonam.authzmanager.controller.admin.user;
 
+import me.sonam.authzmanager.AuthzManagerException;
+import me.sonam.authzmanager.controller.util.MessageConstants;
+import me.sonam.authzmanager.controller.util.Util;
 import me.sonam.authzmanager.tokenfilter.TokenService;
-import me.sonam.authzmanager.webclients.AccountWebClient;
-import me.sonam.authzmanager.webclients.OauthClientWebClient;
-import me.sonam.authzmanager.webclients.UserWebClient;
+import me.sonam.authzmanager.webclients.*;
+import org.apache.tomcat.websocket.AuthenticationException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -26,12 +28,16 @@ public class DeleteMyDataController {
     private final TokenService tokenService;
     private final OauthClientWebClient oauthClientWebClient;
     private final UserWebClient userWebClient;
+    private final SettingWebClient settingWebClient;
+    private final RoleWebClient roleWebClient;
 
     public DeleteMyDataController(UserWebClient userWebClient, OauthClientWebClient oauthClientWebClient,
-                                  TokenService tokenService) {
+                                  TokenService tokenService, SettingWebClient settingWebClient, RoleWebClient roleWebClient) {
         this.userWebClient = userWebClient;
         this.oauthClientWebClient = oauthClientWebClient;
         this.tokenService = tokenService;
+        this.settingWebClient = settingWebClient;
+        this.roleWebClient = roleWebClient;
     }
 
     @GetMapping
@@ -47,11 +53,21 @@ public class DeleteMyDataController {
         String userIdString = defaultOidcUser.getAttribute("userId");
 
         LOG.info("delete user data for userId: {}", userIdString);
-        String accessToken = tokenService.getAccessToken();
+        final String accessToken = tokenService.getAccessToken();
         LOG.info("accessToken {}", accessToken);
+        UUID userId = Util.getLoggedInUserId();
 
-        return oauthClientWebClient.deleteClient(accessToken).then(
-                userWebClient.deleteUser(accessToken)).thenReturn(PATH);
+        return settingWebClient.getDefaultOrganization(accessToken, userId)
+                .switchIfEmpty(Mono.error(new AuthzManagerException("no default organization found")))
+                .flatMap(orgId -> roleWebClient.isSuperAdminInOrgId(accessToken, userId, orgId).zipWith(Mono.just(orgId)))
+                .flatMap(objects -> {
+                    if (!objects.getT1()) {
+                        model.addAttribute("error", MessageConstants.NOT_SUPERADMIN + " " + objects.getT2());
+                        return Mono.error(new AuthenticationException(MessageConstants.NOT_SUPERADMIN));
+                    }
+                    return oauthClientWebClient.deleteClient(accessToken).then(
+                            userWebClient.deleteUser(accessToken, objects.getT2())).thenReturn(PATH);
+                });
     }
 
 }
